feat: Complete logging into an existing account section
Logging into an existing account is similar to registering a new account,
so the section is very sparse and references the previous section.

This commit also introduces cross-references (named anchor) in markdown,
which might not work with all markdown renderers.

The way to do it was refernced from this link:
https://stackoverflow.com/questions/5319754/cross-reference-named-anchor-in-markdown/7335259#7335259